fakeNewsDetection/requirements.txt
bert-extractive-summarizer
blis
boto3
botocore
catalogue
certifi
chardet
click
cymem
filelock
huggingface-hub
idna
imbalanced-learn
imblearn
Jinja2
jmespath
joblib
MarkupSafe
murmurhash
numpy
neuralcoref
packaging
pandas
pathy
preshed
pydantic
pyparsing
python-dateutil
pytz
PyYAML
regex
requests
s3transfer
sacremoses
scikit-learn
scipy
six
sklearn
smart-open
spacy
spacy-legacy
srsly
thinc
threadpoolctl
tokenizers
torch
tqdm
transformers
typer
typing-extensions
urllib3
wasabi